+++ xp/xp_test.go
@@ -1,11 +1,13 @@
 package xp
 
 import (
+	"fmt"
 	"reflect"
 	"sync"
 	"testing"
 
 	p "git.sr.ht/~jrswab/akordo/plugins"
+	"git.sr.ht/~jrswab/akordo/xp/mocks"
 	dg "github.com/bwmarrin/discordgo"
 )
 
@@ -31,6 +33,7 @@ func TestNewXpStore(t *testing.T) {
 				dgs:         testSession,
 				callRec:     p.NewRecorder(),
 				defaultFile: DefaultFile,
+				tiers:       &autoRanks{Tiers: make(map[string]float64)},
 			},
 		},
 	}
@@ -159,6 +162,147 @@ func TestSystem_ManipulateXP(t *testing.T) {
 			x.ManipulateXP(tt.args.action, tt.args.msg)
 			if data.Users[tt.args.msg.Author.ID] != tt.want {
 				t.Errorf("got %.2f, want %.2f", data.Users[tt.args.msg.Author.ID], tt.want)
+			}
+		})
+	}
+}
+
+func TestSystem_AutoPromote(t *testing.T) {
+	testTiers := make(map[string]float64)
+	testTiers["Crew"] = 100
+	testTiers["Ensign"] = 200
+
+	testUsers := make(map[string]float64)
+	testUsers["1111"] = 100
+	testUsers["2222"] = 10
+	testUsers["4444"] = 50
+
+	role1 := &dg.Role{ID: "0000", Name: "Crew"}
+	role2 := &dg.Role{ID: "1111", Name: "Ensign"}
+	roleSlice := []*dg.Role{role1, role2}
+
+	mockSess := new(mocks.AkSession)
+	// For "user found and promoted"
+	mockSess.On("GuildRoles", "1111").Return(roleSlice, nil).Once()
+	mockSess.On("GuildMemberRoleAdd", "1111", "1111", "0000").Return(nil).Once()
+
+	// For "bad guild id"
+	mockSess.On("GuildRoles", "0000").Return(nil, fmt.Errorf("some fake error")).Once()
+
+	// For "user not found"
+	mockSess.On("GuildRoles", "1111").Return(roleSlice, nil).Once()
+
+	// For "fail to update user roles"
+	mockSess.On("GuildRoles", "1111").Return(roleSlice, nil).Once()
+	mockSess.On("GuildMemberRoleAdd", "1111", "4444", "").
+		Return(fmt.Errorf("some fake error")).Once()
+
+	type fields struct {
+		data        *xpData
+		tiers       *autoRanks
+		defaultFile string
+		callRec     *p.Record
+		mutex       *sync.Mutex
+		dgs         AkSession
+	}
+	type args struct {
+		msg *dg.MessageCreate
+	}
+	tests := []struct {
+		name    string
+		fields  fields
+		args    args
+		wantErr bool
+	}{
+		{
+			name: "User found and promoted",
+			fields: fields{
+				data:  &xpData{Users: testUsers},
+				tiers: &autoRanks{Tiers: testTiers},
+				dgs:   mockSess,
+			},
+			args: args{
+				msg: &dg.MessageCreate{
+					&dg.Message{
+						GuildID: "1111",
+						Author: &dg.User{
+							ID: "1111",
+						},
+					},
+				},
+			},
+			wantErr: false,
+		},
+		{
+			name: "Bad Guild ID",
+			fields: fields{
+				data:  &xpData{Users: testUsers},
+				tiers: &autoRanks{Tiers: testTiers},
+				dgs:   mockSess,
+			},
+			args: args{
+				msg: &dg.MessageCreate{
+					&dg.Message{
+						GuildID: "0000",
+						Author: &dg.User{
+							ID: "2222",
+						},
+					},
+				},
+			},
+			wantErr: true,
+		},
+		{
+			name: "User not found",
+			fields: fields{
+				data:  &xpData{Users: testUsers},
+				tiers: &autoRanks{Tiers: testTiers},
+				dgs:   mockSess,
+			},
+			args: args{
+				msg: &dg.MessageCreate{
+					&dg.Message{
+						GuildID: "1111",
+						Author: &dg.User{
+							ID: "3333",
+						},
+					},
+				},
+			},
+			wantErr: true,
+		},
+		{
+			name: "Fail to update user roles",
+			fields: fields{
+				data:  &xpData{Users: testUsers},
+				tiers: &autoRanks{Tiers: testTiers},
+				dgs:   mockSess,
+			},
+			args: args{
+				msg: &dg.MessageCreate{
+					&dg.Message{
+						GuildID: "1111",
+						Author: &dg.User{
+							ID: "4444",
+						},
+					},
+				},
+			},
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			x := &System{
+				data:        tt.fields.data,
+				tiers:       tt.fields.tiers,
+				defaultFile: tt.fields.defaultFile,
+				callRec:     tt.fields.callRec,
+				mutex:       tt.fields.mutex,
+				dgs:         tt.fields.dgs,
+			}
+			if err := x.AutoPromote(tt.args.msg); (err != nil) != tt.wantErr {
+				t.Errorf("System.AutoPromote() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
 	}